Refactor : 영화관 좌석들에 일급 컬렉션 적용
영화관 좌석 의미하는 List<List<Seat>>를 CinemaSeats로 만듬
